+++ packages/expo-notifications/ios/EXNotifications/Notifications/Records.swift
@@ -212,10 +212,11 @@ struct CategoryTextInputActionRecord: Record {
     self.submitButtonTitle = textInputAction.textInputButtonTitle
   }
 
-  func toUNTextInputNotificationAction(identifier: String, title: String) -> UNTextInputNotificationAction {
+  func toUNTextInputNotificationAction(identifier: String, title: String, options: UNNotificationActionOptions) -> UNTextInputNotificationAction {
     return UNTextInputNotificationAction(
       identifier: identifier,
       title: title,
+      options: options,
       textInputButtonTitle: submitButtonTitle ?? "",
       textInputPlaceholder: placeholder ?? ""
     )
@@ -265,9 +266,7 @@ public struct CategoryActionRecord: Record {
       let buttonTitle = buttonTitle else {
       return nil
     }
-    if let textInput = textInput {
-      return textInput.toUNTextInputNotificationAction(identifier: identifier, title: buttonTitle)
-    }
+
     var notificationOptions: UNNotificationActionOptions = []
     if let optionsParams = options {
       if optionsParams.opensAppToForeground == true {
@@ -279,6 +278,9 @@ public struct CategoryActionRecord: Record {
       if optionsParams.isAuthenticationRequired == true {
         notificationOptions.insert(.authenticationRequired)
       }
+    }
+    if let textInput = textInput {
+      return textInput.toUNTextInputNotificationAction(identifier: identifier, title: buttonTitle, options: notificationOptions)
     }
     return UNNotificationAction(identifier: identifier, title: buttonTitle, options: notificationOptions)
   }
@@ -368,9 +370,9 @@ public struct CategoryRecord: Record {
   }
 
   func toUNNotificationCategory() -> UNNotificationCategory {
-    let intentIdentifiers: [String] = options?.intentIdentifiers as? [String] ?? []
-    let previewPlaceholder: String? = options?.previewPlaceholder as? String
-    let categorySummaryFormat: String? = options?.categorySummaryFormat as? String
+    let intentIdentifiers: [String] = options?.intentIdentifiers ?? []
+    let previewPlaceholder: String? = options?.previewPlaceholder
+    let categorySummaryFormat: String? = options?.categorySummaryFormat
     let actionsArray = actions?.compactMap { action in
       return action.toUNNotificationAction()
     } ?? []